Rewrote the entire bot using Arikawa, added some functionalities etc.

Check full commit history of the branch to see what has changed

+++ query/character.go
@@ -22,7 +22,7 @@ type CharStruct struct {
 
 // CharacterStruct represent character object
 type CharacterStruct struct {
-	ID      int64  `json:"id"`
+	ID      uint   `json:"id"`
 	SiteURL string `json:"siteUrl"`
 	Image   struct {
 		Large string `json:"large"`
@@ -75,7 +75,6 @@ func CharSearch(input CharSearchInput) (response CharSearchStruct, err error) {
 	} else {
 		req.Var("name", input.Name)
 	}
-
 	err = graphql.NewClient(graphURL).Run(context.Background(), req, &response)
 	return
 }
